+++ packages/vault-connector-entity-storage/tests/entityStorageVaultConnector.spec.ts
@@ -1,10 +1,10 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { Converter } from "@gtsc/core";
-import { MemoryEntityStorageConnector } from "@gtsc/entity-storage-connector-memory";
-import { EntityStorageConnectorFactory } from "@gtsc/entity-storage-models";
-import { nameof } from "@gtsc/nameof";
-import { VaultEncryptionType, VaultKeyType } from "@gtsc/vault-models";
+import { Converter } from "@twin.org/core";
+import { MemoryEntityStorageConnector } from "@twin.org/entity-storage-connector-memory";
+import { EntityStorageConnectorFactory } from "@twin.org/entity-storage-models";
+import { nameof } from "@twin.org/nameof";
+import { VaultEncryptionType, VaultKeyType } from "@twin.org/vault-models";
 import type { VaultKey } from "../src/entities/vaultKey";
 import type { VaultSecret } from "../src/entities/vaultSecret";
 import { EntityStorageVaultConnector } from "../src/entityStorageVaultConnector";