+++ packages/vault-connector-entity-storage/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/vault-connector-entity-storage - Changelog
+# @twin.org/vault-connector-entity-storage - Changelog
 
-## v0.0.3
+## v0.0.1-next.2
 
 - Initial Release