+++ packages/vault-models/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/vault-models - Changelog
+# @twin.org/vault-models - Changelog
 
-## v0.0.3
+## v0.0.1-next.2
 
 - Initial Release